Add note on SAVE(/fork) and used memory
Modern Linuses are able to fork with copy-on-write pages. Unfortunately,
"Linux can't tell in advance how much memory the child will take, so if
the overcommit_memory setting is set to zero fork will fail unless there
is as much free RAM as required to really duplicate all the parent
memory pages".
If the server is dedicated to Redis, it's safe to set overcommit_memory
to 1, otherwise, I don't recommand it (you really don't know what side
effects could be, unless you're a sysadmin).